move build-tagged shardwidth files to subpackage
