+++ shardwidth/16.go
@@ -1,0 +1,5 @@
+// +build shardwidth16
+
+package shardwidth
+
+const Exponent = 16
