+++ shardwidth/17.go
@@ -1,0 +1,5 @@
+// +build shardwidth17
+
+package shardwidth
+
+const Exponent = 17
