+++ shardwidth/18.go
@@ -1,0 +1,5 @@
+// +build shardwidth18
+
+package shardwidth
+
+const Exponent = 18
