+++ shardwidth/19.go
@@ -1,0 +1,5 @@
+// +build shardwidth19
+
+package shardwidth
+
+const Exponent = 19
