+++ shardwidth/20.go
@@ -1,5 +1,5 @@
 // +build !shardwidth16,!shardwidth17,!shardwidth18,!shardwidth19,!shardwidth21,!shardwidth22,!shardwidth23,!shardwidth24,!shardwidth25,!shardwidth26,!shardwidth27,!shardwidth28,!shardwidth29,!shardwidth30,!shardwidth31,!shardwidth32
 
-package pilosa
+package shardwidth
 
-const shardWidthExponent = 20
+const Exponent = 20
