+++ shardwidth/21.go
@@ -1,0 +1,5 @@
+// +build shardwidth21
+
+package shardwidth
+
+const Exponent = 21
