+++ shardwidth/22.go
@@ -1,0 +1,5 @@
+// +build shardwidth22
+
+package shardwidth
+
+const Exponent = 22
