+++ shardwidth/23.go
@@ -1,0 +1,5 @@
+// +build shardwidth23
+
+package shardwidth
+
+const Exponent = 23
