+++ shardwidth/24.go
@@ -1,0 +1,5 @@
+// +build shardwidth24
+
+package shardwidth
+
+const Exponent = 24
