+++ shardwidth/25.go
@@ -1,0 +1,5 @@
+// +build shardwidth25
+
+package shardwidth
+
+const Exponent = 25
