+++ shardwidth/26.go
@@ -1,0 +1,5 @@
+// +build shardwidth26
+
+package shardwidth
+
+const Exponent = 26
