+++ shardwidth/27.go
@@ -1,0 +1,5 @@
+// +build shardwidth27
+
+package shardwidth
+
+const Exponent = 27
