+++ shardwidth/28.go
@@ -1,0 +1,5 @@
+// +build shardwidth28
+
+package shardwidth
+
+const Exponent = 28
