+++ shardwidth/29.go
@@ -1,0 +1,5 @@
+// +build shardwidth29
+
+package shardwidth
+
+const Exponent = 29
