+++ shardwidth/30.go
@@ -1,0 +1,5 @@
+// +build shardwidth30
+
+package shardwidth
+
+const Exponent = 30
