+++ shardwidth/31.go
@@ -1,0 +1,5 @@
+// +build shardwidth31
+
+package shardwidth
+
+const Exponent = 31
