+++ shardwidth/32.go
@@ -1,0 +1,5 @@
+// +build shardwidth32
+
+package shardwidth
+
+const Exponent = 32
