--- shardwidth22.go
@@ -1,5 +1,0 @@
-// +build shardwidth22
-
-package pilosa
-
-const shardWidthExponent = 22
